add exception handelar
