+++ src/main/java/com/weather/application/http/client/HttpDailyForecastRequest.java
@@ -1,7 +1,9 @@
 package com.weather.application.http.client;
 
+import com.vaadin.flow.server.VaadinSession;
 import com.weather.application.data.dto.DailyForecast;
 import com.fasterxml.jackson.databind.ObjectMapper;
+import com.weather.application.exception.CustomErrorHandler;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
