+++ src/main/java/com/weather/application/views/DailyForecastView.java
@@ -2,8 +2,8 @@ package com.weather.application.views;
 
 
 import com.weather.application.data.entity.Favourites;
-import com.weather.application.data.service.FavouritesService;
-import com.weather.application.data.service.WaService;
+import com.weather.application.service.FavouritesService;
+import com.weather.application.service.WaService;
 import com.weather.application.data.dto.DailyForecast;
 import com.weather.application.data.dto.GeoCode;
 import com.weather.application.data.dto.HourlyForecast;
